Test.Hspec.WithTempFile

Description

Support for easily running golden tests using temporary files

Documentation

goldenWith

Arguments

:: Eq golden

=> OsPath
base dir containing golden files

-> Golden golden actual
the test specification

-> actual
the actual output

-> Spec

Run a golden test, where the golden files are all determined based on the name of the test, and with respect to the given base dir.

goldenTest :: Eq golden => Golden golden actual -> actual -> Spec

Runs a golden test exaclty as specified.

data Golden golden actual

A specification of a golden test. In particular, a test where values of type actual are seriealized into something of type golden, and then written to a file. The content of this file is then compared with the correct answer that is deserialized from a goldenFile.

Constructors

Golden

Fields

name :: OsString

description of the test name

actualWriter :: ActualWriter actual golden

how to write the actual file output onto disk

writeGolden :: OsPath -> golden -> IO ()

how to write the golden file

goldenFile :: OsPath

how to write the actual test output to the given file

readGoldenFile :: OsPath -> IO golden

the filePath that stores the golden output

goldenFilePolicy :: GoldenFilePolicy

what to do with the golden file, in particular if it is missing.

actualFile :: ActualFile

file where/how to store the test output

actualFilePolicy :: ActualFilePolicy

What to do with the actual file

prettyActual :: actual -> String

in case the test fails, how to show the actual input

prettyGolden :: golden -> String

in case the test fails, how to show the difference between the exected and actual outputs.

allowFail :: AllowFail

Allow the test to (temporarily?) fail

byteStringGolden :: Show testOutput => Golden ByteString testOutput

The default way of showing the result of golden tests.

The main scheme is to convert your testOutput into a ByteString, and write this bytestring to a temporary file. The expected test output (stored at goldenFile) is then read, and the output that is written to the temporary file is compared with the golden file. If the output matches the temporary file is deleted. If the test fails, the output is kept.

textGolden :: Show testOutput => Golden Text testOutput

Same as byteStringGolden, except that to compare the file contents it reads and writes the data as UTF-8 encoded text.

dimapWith

Arguments

:: (OsPath -> golden' -> IO ())
the new writeGolden implementation

-> (golden' -> String)
the new prettyDiff function

-> (actual' -> actual)

-> (golden -> golden')

-> Golden golden actual

-> Golden golden' actual'

Convenience method to create test specifications from other ones.

mapGolden :: (golden -> golden') -> (golden' -> golden) -> Golden golden actual -> Golden golden' actual

Mapping function to transform the golden

data GoldenFilePolicy

What to do when the golden file is missing

Constructors

FailOnMissing

CopyWithSuffix
copies the actual file to 'nameOfTheGoldenFile.actual' | AskInteractively

data ActualWriter actual golden

How to write a test to the golden file.

Constructors

Convert (actual -> golden)
convert the actual into a value of type golden, so that we can write it to a file using writeGolden

WriteActual (OsPath -> actual -> IO ())
directly write the actual output to a file.

data ActualFile

Data type describing where to write the file to. Either to a temporary file or a reglar one with a specific file path.

Constructors

TempFile

Fields

tempDir :: Maybe OsPath

use the system default if Nothing

PermanentFile OsPath
absolute path to a normal file

fromTestName :: OsString -> OsPath

turn the test name into a somewhat reasonable fileName

data ActualFilePolicy

What to do with the actual test output file.

Constructors

Discard
always discard the actual file, no matter what the test outcome

KeepOnFailure
only keep the output file on failure

KeepAlways
always (i.e. independent of the test output) keep it

data Diff a

Specification of the expected output together with the received output.

Constructors

Diff

Fields

expected :: a

actual :: a

rawGoldenTest :: Eq golden => GoldenTest golden actual -> Spec

Combinator to run a golden test as is.

data GoldenTest golden actual

A Golden test specification together with a particular output

Constructors

GoldenTest

Fields

testSpec :: Golden golden actual

the specification of te test

testOutput :: actual

the output of the test

data AllowFail

Specify whether a test is allowed to fail

Constructors

AllowFail

RequireSuccess